refactor: ipc/network
