+++ packages/bruno-electron/src/utils/auth.js
@@ -1,28 +1,49 @@
-const crypto = require('crypto');
-const { URL } = require('url');
+const { fromIni } = require('@aws-sdk/credential-providers');
+const { aws4Interceptor } = require('aws4-axios');
 
-function isStrPresent(str) {
-  return str && str.trim() !== '' && str.trim() !== 'undefined';
+async function resolveAwsV4Credentials(request) {
+  const awsv4 = request.awsv4config;
+  if (isStrPresent(awsv4.profileName)) {
+    try {
+      credentialsProvider = fromIni({
+        profile: awsv4.profileName
+      });
+      credentials = await credentialsProvider();
+      awsv4.accessKeyId = credentials.accessKeyId;
+      awsv4.secretAccessKey = credentials.secretAccessKey;
+      awsv4.sessionToken = credentials.sessionToken;
+    } catch {
+      console.error('Failed to fetch credentials from AWS profile.');
+    }
+  }
+  return awsv4;
 }
 
-function stripQuotes(str) {
-  return str.replace(/"/g, '');
-}
+function addAwsV4Interceptor(axiosInstance, request) {
+  if (!request.awsv4config) {
+    console.warn('No Auth Config found!');
+    return;
+  }
 
-function containsDigestHeader(response) {
-  const authHeader = response?.headers?.['www-authenticate'];
-  return authHeader ? authHeader.trim().toLowerCase().startsWith('digest') : false;
-}
+  const awsv4 = request.awsv4config;
+  if (!isStrPresent(awsv4.accessKeyId) || !isStrPresent(awsv4.secretAccessKey)) {
+    console.warn('Required Auth Fields are not present');
+    return;
+  }
 
-function containsAuthorizationHeader(originalRequest) {
-  return Boolean(
-    originalRequest.headers['Authorization'] ||
-    originalRequest.headers['authorization']
-  );
-}
+  const interceptor = aws4Interceptor({
+    options: {
+      region: awsv4.region,
+      service: awsv4.service
+    },
+    credentials: {
+      accessKeyId: awsv4.accessKeyId,
+      secretAccessKey: awsv4.secretAccessKey,
+      sessionToken: awsv4.sessionToken
+    }
+  });
 
-function md5(input) {
-  return crypto.createHash('md5').update(input).digest('hex');
+  axiosInstance.interceptors.request.use(interceptor);
 }
 
 function addDigestInterceptor(axiosInstance, request) {
@@ -123,4 +144,32 @@ function addDigestInterceptor(axiosInstance, request) {
   );
 }
 
-module.exports = { addDigestInterceptor };
+function containsDigestHeader(response) {
+  const authHeader = response?.headers?.['www-authenticate'];
+  return authHeader ? authHeader.trim().toLowerCase().startsWith('digest') : false;
+}
+
+function containsAuthorizationHeader(originalRequest) {
+  return Boolean(
+    originalRequest.headers['Authorization'] ||
+    originalRequest.headers['authorization']
+  );
+}
+
+function md5(input) {
+  return crypto.createHash('md5').update(input).digest('hex');
+}
+
+function isStrPresent(str) {
+  return str && str !== '' && str !== 'undefined';
+}
+
+function stripQuotes(str) {
+  return str.replace(/"/g, '');
+}
+
+module.exports = {
+  addAwsV4Interceptor,
+  resolveAwsV4Credentials,
+  addDigestInterceptor
+}
